+++ Common/src/main/java/fuzs/puzzleslib/api/client/init/v1/ModelLayerFactory.java
@@ -1,62 +1,53 @@
 package fuzs.puzzleslib.api.client.init.v1;
 
+import fuzs.puzzleslib.impl.client.init.ModelLayerFactoryImpl;
 import net.minecraft.client.model.geom.ModelLayerLocation;
-import net.minecraft.resources.ResourceLocation;
 
 /**
- * helper class for creating {@link ModelLayerFactory} objects with a provided namespace
+ * A helper class for creating {@link ModelLayerFactoryImpl} instances with a provided namespace.
  */
-public final class ModelLayerFactory {
-    /**
-     * namespace used for {@link ResourceLocation}
-     */
-    private final String namespace;
+public interface ModelLayerFactory {
 
     /**
-     * @param namespace registry namespace
-     */
-    private ModelLayerFactory(String namespace) {
-        this.namespace = namespace;
-    }
-
-    /**
-     * @param path entity name
-     * @return location for main
-     */
-    public ModelLayerLocation register(String path) {
-        return this.register(path, "main");
-    }
-
-    /**
-     * @param path entity name
-     * @param layer layer name
-     * @return location for <code>layer</code>
-     */
-    public ModelLayerLocation register(String path, String layer) {
-        return new ModelLayerLocation(new ResourceLocation(this.namespace, path), layer);
-    }
-
-    /**
-     * @param path entity name
-     * @return location for inner armor
-     */
-    public ModelLayerLocation registerInnerArmor(String path) {
-        return this.register(path, "inner_armor");
-    }
-
-    /**
-     * @param path entity name
-     * @return location for outer armor
-     */
-    public ModelLayerLocation registerOuterArmor(String path) {
-        return this.register(path, "outer_armor");
-    }
-
-    /**
+     * Creates a new instance from a namespace.
+     *
      * @param namespace registry namespace
      * @return new registry from <code>namespace</code>
      */
-    public static ModelLayerFactory of(String namespace) {
-        return new ModelLayerFactory(namespace);
+    static ModelLayerFactory from(String namespace) {
+        return new ModelLayerFactoryImpl(namespace);
     }
+
+    /**
+     * Creates a new {@link ModelLayerLocation}, the used layer is <code>main</code>.
+     *
+     * @param path entity name
+     * @return location for main
+     */
+    ModelLayerLocation register(String path);
+
+    /**
+     * Creates a new {@link ModelLayerLocation}, the used layer is provided by the parameter <code>layer</code>.
+     *
+     * @param path  entity name
+     * @param layer layer name
+     * @return location for <code>layer</code>
+     */
+    ModelLayerLocation register(String path, String layer);
+
+    /**
+     * Creates a new {@link ModelLayerLocation}, the used layer is <code>inner_armor</code>.
+     *
+     * @param path entity name
+     * @return location for inner armor
+     */
+    ModelLayerLocation registerInnerArmor(String path);
+
+    /**
+     * Creates a new {@link ModelLayerLocation}, the used layer is <code>outer_armor</code>.
+     *
+     * @param path entity name
+     * @return location for outer armor
+     */
+    ModelLayerLocation registerOuterArmor(String path);
 }